Refactored some old v5 code, to get around OCC native collections can't be used in managed

+++ Xbim.Geometry.Engine/XbimEdge.cpp
@@ -799,7 +799,7 @@ namespace Xbim
 		int XbimEdge::GetHashCode()
 		{
 			if (!IsValid) return 0;
-			return std::hash<TopoDS_Edge>()(*pEdge);
+			return (int)std::hash<TopoDS_Edge>()(*pEdge);
 		}
 
 		bool XbimEdge::operator ==(XbimEdge^ left, XbimEdge^ right)
